fix: move awaitCurrentOffering/awaitOffering tests to testDefaults source set
